Refactor: inputFormat 함수 utils로 분리 Refactor: toLocaleString unicode 명시
